Starting to be able to create a load balancer.
I started by adding a 'NewLoadbalancer' model .. but now I think,
because so many bits are optional, I'll leave it how it was and just
pass all the parameters to a function.

I'll drop the 'validation' layer because the server does pretty good
validation and returns a good reply.

I'll try to document the way to make a valid call in the API function
itself.